fix: deduplicate a few targets
