+++ .agents/skills/github-cli/references/workflows.md
@@ -301,4 +301,3 @@ gh pr edit <number> --body "$(cat <<'EOF'
 EOF
 )"
 ```
-
